+++ src/lib/components/directive-collection.ts
@@ -5,6 +5,7 @@ import { FixedHeaderDirective } from './tables/fixed-header/fixed-header.directi
 import { TerraLabelTooltipDirective } from '../helpers/terra-label-tooltip.directive';
 import { FormEntryContainerDirective } from './forms/form/form-entry/form-entry-container.directive';
 import { CKEditorDirective } from './editors/ck-editor/ck-editor.directive';
+import { TooltipDirective } from './tooltip/tooltip.directive';
 import { FloatTheadDirective } from './tables/data-table/float-thead/float-thead.directive';
 
 export const directives:Array<Type<any>> = [
@@ -14,5 +15,6 @@ export const directives:Array<Type<any>> = [
     TerraLabelTooltipDirective,
     FormEntryContainerDirective,
     CKEditorDirective,
+    TooltipDirective,
     FloatTheadDirective
 ];