+++ src/lib/components/tables/data-table/interfaces/terra-data-table-header-cell.interface.ts
@@ -13,6 +13,9 @@ export interface TerraDataTableHeaderCellInterface
     sortBy?:string;
     isHidden?:boolean;
     tooltipText?:string;
+    /**
+     * @deprecated since v4. Is replaced by the TooltipDirective and will be removed with the next major version.
+     */
     tooltipPlacement?:string;
     textAlign?:TerraTextAlignEnum;
 }