+++ src/lib/components/tables/group-function/terra-group-function.component.spec.ts
@@ -7,9 +7,9 @@ import { LocalizationModule } from 'angular-l10n';
 import { l10nConfig } from '../../../../app/translation/l10n.config';
 import { TerraGroupFunctionComponent } from './terra-group-function.component';
 import { TerraButtonComponent } from '../../buttons/button/terra-button.component';
-import { TooltipModule } from 'ngx-bootstrap';
 import { HttpClientModule } from '@angular/common/http';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { TooltipDirective } from '../../tooltip/tooltip.directive';
 
 describe('Component: TerraGroupFunctionComponent', () =>
 {
@@ -19,12 +19,11 @@ describe('Component: TerraGroupFunctionComponent', () =>
     beforeEach(async(() =>
     {
         TestBed.configureTestingModule({
-            declarations: [
-                TerraGroupFunctionComponent,
-                TerraButtonComponent
+            declarations: [TooltipDirective,
+                           TerraGroupFunctionComponent,
+                           TerraButtonComponent
             ],
             imports:      [
-                TooltipModule.forRoot(),
                 NoopAnimationsModule,
                 HttpClientModule,
                 LocalizationModule.forRoot(l10nConfig)